Notes
This can be used to check that formulae have been copied correctly and to pick up any hardcoded cells. It can also speed up checking by showing which cells have original formulae. Once these are checked, then you have only to check they have been copied correctly

A copy is made of the current sheet, and shading added to each cell with a formula
The shading shows if formulae are copied from the left (horizontal shading) or above (vertical shading), both (crosshatch shading), or neither (solid shading)
It takes some time because for each cell with a formula, it literally copies the formulae from left and above into that cell, and checks if they are identical