Notes
This code makes it easy for you to add centralised comments when reviewing an Excel workbook, with links to the cells you are referencing
1. It adds a Checks sheet to a workbook (if it doesn't exist already)
2. It adds a new row in that sheet, starting with a link to the cell you were in when you ran the code
3. You can then add your comments

Yes, you can add comments to individual cells, but will people see them and how do they respond to them? It's better to centralise comments when reviewing a workbook.